+++ libs/base/thread_util.h
@@ -38,7 +38,7 @@ namespace thread
 /// std::thread::hardware_concurrency() does not always return a
 /// useable value on all systems. This attempts to remedy that by
 /// trying that first, then using other older API to guess.
-int core_count( void );
+long core_count( void );
 
 /// @brief Override core count returned by above core_count
 ///
@@ -46,7 +46,7 @@ int core_count( void );
 /// threads for a thread pool or similar object will instead return
 /// the count provided here. Any value less than 0 will reset this,
 /// and the core_count will return to returning the system count.
-void override_core_count( int cnt = -1 );
+void override_core_count( long cnt = -1 );
 
 } // namespace thread
